move update_status_from_rq to model
